Fixed integration on multiple points. This deserves an explanation
So, thing is, I didn't notice I was making the tests for the PDF
on high values for x, e.g, x = 10, 20. I didn't notice that, when
x was lower, the PDF was completely messed up, with relative errors
greater than one.

How did I find where was the causing error? After a lot of searching, I
started messing with the code without any purpose and ended up putting
"&& 0" on the condition for the reduction of the subinterval. And, just
like magic, the graph generated with stable_plot fixed itself and the
PDF results were good.

So, this is how I noticed that the subinterval length was erroneous.

Funnily enough, with just 1 subinterval the results were pretty good.